+++ src/generator.rs
@@ -2,15 +2,21 @@
 
 use crate::{Scru128Id, MAX_COUNTER_HI, MAX_COUNTER_LO, MAX_TIMESTAMP};
 
-#[cfg(feature = "std")]
-pub use default_rng::DefaultRng;
+/// A trait that defines the minimum random number generator interface for [`Scru128Generator`].
+pub trait Scru128Rng {
+    /// Return the next random `u32`.
+    fn next_u32(&mut self) -> u32;
+}
 
-/// The default random number generator used by [`Scru128Generator`].
-///
-/// No default random number generator is available in `no_std` mode.
-#[cfg(not(feature = "std"))]
-#[derive(Clone, Debug)]
-pub struct DefaultRng(());
+#[cfg(feature = "rand")]
+#[cfg_attr(docsrs, doc(cfg(feature = "rand")))]
+impl<T: rand::RngCore> Scru128Rng for T {
+    fn next_u32(&mut self) -> u32 {
+        self.next_u32()
+    }
+}
+
+pub use default_rng::DefaultRng;
 
 /// Represents a SCRU128 ID generator that encapsulates the monotonic counters and other internal
 /// states.
@@ -18,7 +24,7 @@ pub struct DefaultRng(());
 /// # Examples
 ///
 /// ```rust
-/// # #[cfg(feature = "std")]
+/// # #[cfg(feature = "default_rng")]
 /// # {
 /// use scru128::Scru128Generator;
 ///
@@ -33,7 +39,7 @@ pub struct DefaultRng(());
 /// synchronization mechanisms to control the scope of guaranteed monotonicity:
 ///
 /// ```rust
-/// # #[cfg(feature = "std")]
+/// # #[cfg(feature = "default_rng")]
 /// # {
 /// use scru128::Scru128Generator;
 /// use std::sync::{Arc, Mutex};
@@ -93,14 +99,14 @@ pub struct Scru128Generator<R = DefaultRng> {
     rng: R,
 }
 
-impl<R: rand::RngCore> Scru128Generator<R> {
+impl<R: Scru128Rng> Scru128Generator<R> {
     /// Creates a generator object with a specified random number generator. The specified random
     /// number generator should be cryptographically strong and securely seeded.
     ///
     /// # Examples
     ///
     /// ```rust
-    /// # #[cfg(feature = "std")]
+    /// # #[cfg(feature = "default_rng")]
     /// # {
     /// use scru128::Scru128Generator;
     ///
@@ -198,21 +204,23 @@ impl<R: rand::RngCore> Scru128Generator<R> {
     }
 }
 
+#[cfg(any(feature = "default_rng", test))]
+#[cfg_attr(docsrs, doc(cfg(feature = "default_rng")))]
+impl Scru128Generator {
+    /// Creates a generator object with the default random number generator.
+    pub fn new() -> Self {
+        Default::default()
+    }
+}
+
 #[cfg(feature = "std")]
 #[cfg_attr(docsrs, doc(cfg(feature = "std")))]
 mod std_ext {
-    use super::{Scru128Generator, Scru128Id};
+    use super::{Scru128Generator, Scru128Id, Scru128Rng};
     use std::{iter, time};
 
     /// The default timestamp rollback allowance.
     const DEFAULT_ROLLBACK_ALLOWANCE: u64 = 10_000; // 10 seconds
-
-    impl Scru128Generator {
-        /// Creates a generator object with the default random number generator.
-        pub fn new() -> Self {
-            Default::default()
-        }
-    }
 
     /// Returns the current Unix timestamp in milliseconds.
     fn unix_ts_ms() -> u64 {
@@ -222,7 +230,7 @@ mod std_ext {
             .as_millis() as u64
     }
 
-    impl<R: rand::RngCore> Scru128Generator<R> {
+    impl<R: Scru128Rng> Scru128Generator<R> {
         /// Generates a new SCRU128 ID object from the current `timestamp`, or resets the generator
         /// upon significant timestamp rollback.
         ///
@@ -239,6 +247,8 @@ mod std_ext {
         /// # Examples
         ///
         /// ```rust
+        /// # #[cfg(feature = "default_rng")]
+        /// # {
         /// use scru128::Scru128Generator;
         ///
         /// let mut g = Scru128Generator::new();
@@ -247,6 +257,7 @@ mod std_ext {
         ///     .generate_or_abort()
         ///     .expect("The clock went backwards by ten seconds!");
         /// assert!(x < y);
+        /// # }
         /// ```
         pub fn generate_or_abort(&mut self) -> Option<Scru128Id> {
             self.generate_or_abort_core(unix_ts_ms(), DEFAULT_ROLLBACK_ALLOWANCE)
@@ -259,14 +270,17 @@ mod std_ext {
     /// # Examples
     ///
     /// ```rust
+    /// # #[cfg(feature = "default_rng")]
+    /// # {
     /// use scru128::Scru128Generator;
     ///
     /// let g = Scru128Generator::new();
     /// for (i, e) in g.take(8).enumerate() {
     ///     println!("[{i}] {e}");
     /// }
+    /// # }
     /// ```
-    impl<R: rand::RngCore> Iterator for Scru128Generator<R> {
+    impl<R: Scru128Rng> Iterator for Scru128Generator<R> {
         type Item = Scru128Id;
 
         fn next(&mut self) -> Option<Self::Item> {
@@ -278,7 +292,26 @@ mod std_ext {
         }
     }
 
-    impl<R: rand::RngCore> iter::FusedIterator for Scru128Generator<R> {}
+    impl<R: Scru128Rng> iter::FusedIterator for Scru128Generator<R> {}
+
+    #[cfg(test)]
+    mod tests {
+        /// Is iterable with for-in loop
+        #[test]
+        fn is_iterable_with_for_in_loop() {
+            use super::Scru128Generator;
+
+            let mut i = 0;
+            for e in Scru128Generator::new() {
+                assert!(e.timestamp() > 0);
+                i += 1;
+                if i > 100 {
+                    break;
+                }
+            }
+            assert_eq!(i, 101);
+        }
+    }
 }
 
 #[cfg(test)]
@@ -289,10 +322,7 @@ mod tests_generate_or_reset {
     #[test]
     fn generates_increasing_ids_even_with_decreasing_or_constant_timestamp() {
         let ts = 0x0123_4567_89abu64;
-        #[cfg(feature = "std")]
         let mut g = Scru128Generator::new();
-        #[cfg(not(feature = "std"))]
-        let mut g = Scru128Generator::with_rng(super::tests::no_std_rng());
 
         let mut prev = g.generate_or_reset_core(ts, 10_000);
         assert_eq!(prev.timestamp(), ts);
@@ -309,10 +339,7 @@ mod tests_generate_or_reset {
     #[test]
     fn breaks_increasing_order_of_ids_if_timestamp_goes_backwards_a_lot() {
         let ts = 0x0123_4567_89abu64;
-        #[cfg(feature = "std")]
         let mut g = Scru128Generator::new();
-        #[cfg(not(feature = "std"))]
-        let mut g = Scru128Generator::with_rng(super::tests::no_std_rng());
 
         let mut prev = g.generate_or_reset_core(ts, 10_000);
         assert_eq!(prev.timestamp(), ts);
@@ -335,10 +362,7 @@ mod tests_generate_or_abort {
     #[test]
     fn generates_increasing_ids_even_with_decreasing_or_constant_timestamp() {
         let ts = 0x0123_4567_89abu64;
-        #[cfg(feature = "std")]
         let mut g = Scru128Generator::new();
-        #[cfg(not(feature = "std"))]
-        let mut g = Scru128Generator::with_rng(super::tests::no_std_rng());
 
         let mut prev = g.generate_or_abort_core(ts, 10_000).unwrap();
         assert_eq!(prev.timestamp(), ts);
@@ -355,10 +379,7 @@ mod tests_generate_or_abort {
     #[test]
     fn returns_none_if_timestamp_goes_backwards_a_lot() {
         let ts = 0x0123_4567_89abu64;
-        #[cfg(feature = "std")]
         let mut g = Scru128Generator::new();
-        #[cfg(not(feature = "std"))]
-        let mut g = Scru128Generator::with_rng(super::tests::no_std_rng());
 
         let prev = g.generate_or_abort_core(ts, 10_000).unwrap();
         assert_eq!(prev.timestamp(), ts);
@@ -371,39 +392,12 @@ mod tests_generate_or_abort {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    #[cfg(not(feature = "std"))]
-    pub fn no_std_rng() -> impl rand::RngCore {
-        use rand::SeedableRng as _;
-        let local_var = 0u32;
-        let addr_as_seed = (&local_var as *const u32) as u64;
-        rand_chacha::ChaCha12Rng::seed_from_u64(addr_as_seed)
-    }
-
-    /// Is iterable with for-in loop
-    #[cfg(feature = "std")]
-    #[test]
-    fn is_iterable_with_for_in_loop() {
-        use super::Scru128Generator;
-
-        let mut i = 0;
-        for e in Scru128Generator::new() {
-            assert!(e.timestamp() > 0);
-            i += 1;
-            if i > 100 {
-                break;
-            }
-        }
-        assert_eq!(i, 101);
-    }
-}
-
-#[cfg(feature = "std")]
-#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
 mod default_rng {
-    use rand::{rngs::adapter::ReseedingRng, rngs::OsRng, SeedableRng};
-    use rand_chacha::ChaCha12Core;
+    #[cfg(feature = "default_rng")]
+    use rand::{rngs::adapter::ReseedingRng, rngs::OsRng, SeedableRng as _};
+
+    #[cfg(all(not(feature = "default_rng"), test))]
+    use rand::{rngs::StdRng, SeedableRng as _};
 
     /// The default random number generator used by [`Scru128Generator`].
     ///
@@ -412,48 +406,59 @@ mod default_rng {
     /// the same strategy as that employed by [`ThreadRng`]; see the docs of `rand` crate for a
     /// detailed discussion on the strategy.
     ///
-    /// This structure does exist under the `no_std` environment but is not able to be instantiated
-    /// or used as a random number generator.
+    /// This structure does exist without the `default_rng` feature flag but is not able to be
+    /// instantiated or used as a random number generator.
     ///
     /// [`Scru128Generator`]: super::Scru128Generator
     /// [`ChaCha12Core`]: rand_chacha::ChaCha12Core
     /// [`OsRng`]: rand::rngs::OsRng
     /// [`ReseedingRng`]: rand::rngs::adapter::ReseedingRng
     /// [`ThreadRng`]: https://docs.rs/rand/0.8/rand/rngs/struct.ThreadRng.html
+    #[cfg_attr(docsrs, doc(cfg(feature = "default_rng")))]
     #[derive(Clone, Debug)]
-    pub struct DefaultRng(ReseedingRng<ChaCha12Core, OsRng>);
+    pub struct DefaultRng {
+        _private: (),
 
+        #[cfg(feature = "default_rng")]
+        inner: ReseedingRng<rand_chacha::ChaCha12Core, OsRng>,
+
+        #[cfg(all(not(feature = "default_rng"), test))]
+        inner: StdRng,
+    }
+
+    #[cfg(any(feature = "default_rng", test))]
+    impl super::Scru128Rng for DefaultRng {
+        fn next_u32(&mut self) -> u32 {
+            rand::RngCore::next_u32(&mut self.inner)
+        }
+    }
+
+    #[cfg(any(feature = "default_rng", test))]
     impl Default for DefaultRng {
         fn default() -> Self {
-            let rng = ChaCha12Core::from_rng(OsRng).expect("could not initialize DefaultRng");
-            Self(ReseedingRng::new(rng, 1024 * 64, OsRng))
+            Self {
+                _private: (),
+
+                #[cfg(feature = "default_rng")]
+                inner: {
+                    let rng = rand_chacha::ChaCha12Core::from_rng(OsRng)
+                        .expect("could not initialize DefaultRng");
+                    ReseedingRng::new(rng, 1024 * 64, OsRng)
+                },
+
+                #[cfg(all(not(feature = "default_rng"), test))]
+                inner: {
+                    let local_var = 0u32;
+                    let addr_as_seed = (&local_var as *const u32) as u64;
+                    StdRng::seed_from_u64(addr_as_seed)
+                },
+            }
         }
     }
-
-    impl rand::RngCore for DefaultRng {
-        fn next_u32(&mut self) -> u32 {
-            self.0.next_u32()
-        }
-
-        fn next_u64(&mut self) -> u64 {
-            self.0.next_u64()
-        }
-
-        fn fill_bytes(&mut self, dest: &mut [u8]) {
-            self.0.fill_bytes(dest)
-        }
-
-        fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), rand::Error> {
-            self.0.try_fill_bytes(dest)
-        }
-    }
-
-    impl rand::CryptoRng for DefaultRng {}
 
     #[cfg(test)]
     mod tests {
-        use super::DefaultRng;
-        use rand::RngCore;
+        use super::{super::Scru128Rng, DefaultRng};
 
         /// Generates unbiased random numbers
         ///