+++ src/lib.rs
@@ -12,7 +12,7 @@
 //! - 80-bit three-layer randomness for global uniqueness
 //!
 //! ```rust
-//! # #[cfg(feature = "std")]
+//! # #[cfg(feature = "global_gen")]
 //! # {
 //! // generate a new identifier object
 //! let x = scru128::new();
@@ -35,10 +35,14 @@
 //!
 //! Default features:
 //!
-//! - `std` enables the primary [`new()`] and [`new_string()`] functions and configures
-//!   [`Scru128Generator`] with the system clock and default random number generator.
-//!   Without `std`, this crate provides limited functionality available under
-//!   `no_std` environments.
+//! - `std` configures [`Scru128Generator`] with the system clock. Without `std`, this
+//!   crate provides limited functionality available under `no_std` environments.
+//! - `rand` enables a blanket implementation for [`rand::RngCore`] to use the
+//!   conforming random number generators with [`Scru128Generator`].
+//! - `default_rng` (implies `std` and `rand`) configures [`Scru128Generator`] with
+//!   the default random number generator.
+//! - `global_gen` (implies `default_rng`) enables the process-wide default global
+//!   generator and the [`new()`] and [`new_string()`] functions.
 //!
 //! Optional features:
 //!
@@ -48,7 +52,7 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
 mod global_gen;
-#[cfg(feature = "std")]
+#[cfg(feature = "global_gen")]
 pub use global_gen::{new, new_string};
 
 mod identifier;