+++ CHANGELOG.md
@@ -27,6 +27,7 @@ a case-insensitive scheme. Other changes include the removal of deprecated APIs.
 
 ### Added
 
+- `global_gen` feature flag to opt in/out default global generator
 - `const` qualifier to `Scru128Id::encode()`
 - `Scru128Id::as_bytes()` and `impl AsRef<[u8]>` for `Scru128Id`